+++ scripts/forks/hddcoin_install.sh
@@ -4,15 +4,18 @@
 #
 
 HDDCOIN_BRANCH=$1
+# On 2021-11-06
+HASH=fa28cc2a12308236706ac122fb6cd98634b840be
 
 if [ -z ${HDDCOIN_BRANCH} ]; then
 	echo 'Skipping HDDCoin install as not requested.'
 else
-	git clone --branch ${HDDCOIN_BRANCH} --recurse-submodules https://github.com/HDDcoin-Network/hddcoin-blockchain.git /hddcoin-blockchain \
-		&& cd /hddcoin-blockchain \
-		&& git submodule update --init mozilla-ca \
-		&& chmod +x install.sh \
-		&& /usr/bin/sh ./install.sh
+	git clone --branch ${HDDCOIN_BRANCH} --recurse-submodules https://github.com/HDDcoin-Network/hddcoin-blockchain.git /hddcoin-blockchain
+	cd /hddcoin-blockchain
+	git submodule update --init mozilla-ca 
+	git checkout $HASH
+	chmod +x install.sh 
+	/usr/bin/sh ./install.sh
 
 	if [ ! -d /chia-blockchain/venv ]; then
 		cd /