+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.4] - 2021-11-?
+ - [Farmr](https://github.com/gilnobrega/farmr) - include `farmr` client for optional monitoring at [farmr.net](https://farmr.net/)
+ - [Stor](https://github.com/Stor-Network/stor-blockchain) - cross-farming support for this blockchain fork.
+ - [Cactus](https://github.com/Cactus-Network/cactus-blockchain) - cross-farming support for this blockchain fork. 
+
 ## [0.6.3] - 2021-11-05
  - [Staicoin](https://github.com/STATION-I/staicoin-blockchain) - cross-farming support for this blockchain fork.
  - Chia - Update to version 1.2.11.  See their [changelog for details](https://github.com/Chia-Network/chia-blockchain/releases/tag/1.2.11).